website/pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["Danis <zinnatullin2012@inbox.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.0"
psycopg2 = "^2.8.6"
celery = "^5.0.5"
django-rosetta = "^0.9.6"
redis = "^3.5.3"
django-crispy-forms = "^1.11.2"
unicodecsv = "^0.14.1"
beautifultable = "^1.0.1"
django-parler = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
